+++ gridcells/training/deepmind/main.py
@@ -7,19 +7,17 @@ from dataclasses import asdict, dataclass
 import torch
 import numpy as np
 import torch.nn as nn
-from PIL import Image
 from tqdm import tqdm
 from torch.utils.data import DataLoader
-import torchvision.transforms as transforms
 from torch.utils.tensorboard import SummaryWriter
 
-from gridcells.validation import sac as SAC
 from gridcells.data import encoder as data_encoder
 from gridcells.models import main as gridcell_models
+from gridcells.data.dataset import EncodedLocationDataset
 from gridcells.validation import views as validation_views
 from gridcells.training.deepmind import epochs as training_epochs
+from gridcells.training.deepmind import reviews as deepmind_reviews
 from gridcells.training.base.rmsprop_tf import RMSprop as RMSprop_tf
-from gridcells.data.dataset import CachedEncodedDataset, EncodedLocationDataset
 
 
 @dataclass
@@ -70,11 +68,12 @@ def train():
 
     date_time = dt.datetime.now().strftime("%m%d_%H%M")
     if device.type == "cuda":
-        run_name = "GPU" + str(torch.cuda.current_device()) + "_DM_" + date_time
+        run_name = "GPU_" + str(torch.cuda.current_device()) + "_DM_" + date_time
     else:
         run_name = "CPU_DM_" + date_time
     print("Current run:", run_name)
 
+    # Display settings in tensorboard/text
     writer = SummaryWriter(f"tmp/tensorboard/{run_name}")
     writer.add_text("config", config.markdown(), 0)
 
@@ -145,7 +144,7 @@ def train():
         # Detailed validation
         if epoch % 10 == 0:
             save_experiment(model, optimizer, config, run_name)
-            ratemaps = make_scored_ratemaps(model, device, test_batch)
+            ratemaps = deepmind_reviews.make_scored_ratemaps(model, device, test_batch)
 
             # Passing a 'sort key' as an argument to the view might be cleaner
             ratemaps = sorted(ratemaps, key=lambda r: -r.s60)
@@ -161,7 +160,7 @@ def train():
                 scores=[r.s90 for r in ratemaps],
             )
             writer.add_figure("validation/s90_ratemaps", fig, epoch)
-            review_path_integration_batch(
+            deepmind_reviews.review_path_integration_batch(
                 model=model,
                 batch=test_batch,
                 device=device,
@@ -189,102 +188,6 @@ def save_experiment(
     torch.save(experiment_state, f"tmp/{run_name}.pt")
 
 
-def review_path_integration_batch(model: nn.Module, batch: dict, device: str, writer, epoch: int, n_place_cells: int):
-    # Draw this many charts
-    n_samples = 10
-    ego_vel = batch["ego_vel"].to(device)
-    encoded_pos = batch["encoded_initial_pos"].to(device)
-    encoded_hd = batch["encoded_initial_hd"].to(device)
-    concat_init = torch.cat([encoded_hd, encoded_pos], axis=2).squeeze()
-    # And for drawing charts
-    init_pos = batch["init_pos"].detach().numpy()
-    target_hd = batch["target_hd"].numpy()
-    target_pos = batch["target_pos"].numpy()
-
-    predicted_positions, predicted_hd, bottlenecks = model(concat_init, ego_vel)
-    predicted_positions = predicted_positions.cpu().detach().numpy()
-    predicted_hd = predicted_hd.cpu().detach().numpy()
-    hd_encoder = data_encoder.DeepMindHeadEncoder()
-    position_encoder = data_encoder.DeepMindPlaceEncoder(n_place_cells)
-
-    for it in range(n_samples):
-        decoded_hd = hd_encoder.decode(predicted_hd[it])
-        # There's a shape incosistency between deepmind pipeline
-        # and the base pipeline, so I need to add a dimension here to re-use plots
-        decoded_hd = decoded_hd[:, np.newaxis]
-        decoded_position = position_encoder.decode(predicted_positions[it])
-
-        fig = validation_views.compare_model_output(
-            init_pos=init_pos[it],
-            target_place=target_pos[it],
-            model_place=decoded_position,
-            target_head=target_hd[it],
-            model_head=decoded_hd,
-        )
-        savepath = f"tmp/dp-review-{it:02}.png"
-        fig.savefig(savepath)
-        image = Image.open(savepath)
-        transform = transforms.Compose([transforms.PILToTensor()])
-        img_tensor = transform(image)
-        writer.add_image(f"path_integration/{it}", img_tensor, epoch)
-
-
-def review_path_integration(model_state_path: str):
-    # Run tests on cpu
-    device = torch.device("cpu")
-
-    # Draw this many charts
-    n_samples = 10
-
-    model = gridcell_models.DeepMindModel()
-    model_state = torch.load(model_state_path)
-    model.load_state_dict(model_state)
-    model.cpu().eval()
-
-    # Make a dataset from a single path that was
-    # not used during trainig and validation
-    paths = glob("data/encoded_pickles/*pickle")
-    path = paths[-2]
-    dataset = CachedEncodedDataset([path])
-    loader = DataLoader(dataset, batch_size=n_samples, shuffle=True)
-
-    hd_encoder = data_encoder.DeepMindHeadEncoder()
-    position_encoder = data_encoder.DeepMindPlaceEncoder()
-
-    # Unpack data for model input
-    batch = next(iter(loader))
-    ego_vel = batch["ego_vel"].to(device)
-    encoded_pos = batch["encoded_initial_pos"].to(device)
-    encoded_hd = batch["encoded_initial_hd"].to(device)
-    concat_init = torch.cat([encoded_hd, encoded_pos], axis=2).squeeze()
-    # And for drawing charts
-    init_pos = batch["init_pos"].detach().numpy()
-    target_hd = batch["target_hd"].numpy()
-    target_pos = batch["target_pos"].numpy()
-
-    predicted_positions, predicted_hd, bottlenecks = model(concat_init, ego_vel)
-    predicted_positions = predicted_positions.detach().numpy()
-    predicted_hd = predicted_hd.detach().numpy()
-
-    for it in range(n_samples):
-        decoded_hd = hd_encoder.decode(predicted_hd[it])
-        # There's a shape incosistency between deepmind pipeline
-        # and the base pipeline, so I need to add a dimension here to re-use plots
-        decoded_hd = decoded_hd[:, np.newaxis]
-        decoded_position = position_encoder.decode(predicted_positions[it])
-
-        fig = validation_views.compare_model_output(
-            init_pos=init_pos[it],
-            target_place=target_pos[it],
-            model_place=decoded_position,
-            target_head=target_hd[it],
-            model_head=decoded_hd,
-        )
-        savepath = f"tmp/dp-review-{it:02}.png"
-        fig.savefig(savepath)
-        return savepath
-
-
 def make_test_batch(path: str, encoder, n_samples: int = 5000) -> dict:
     dataset = EncodedLocationDataset([path], encoder)
     loader = DataLoader(dataset, batch_size=n_samples, shuffle=True)
@@ -292,71 +195,3 @@ def make_test_batch(path: str, encoder, n_samples: int = 5000) -> dict:
     batch = next(iter(loader))
 
     return batch
-
-
-def make_scored_ratemaps(model: nn.Module, device: torch.device, batch: dict) -> list[SAC.Ratemap]:
-    ego_vel = batch["ego_vel"].to(device)
-    encoded_pos = batch["encoded_initial_pos"].to(device)
-    encoded_hd = batch["encoded_initial_hd"].to(device)
-    concat_init = torch.cat([encoded_hd, encoded_pos], axis=2).squeeze()
-    target_pos = batch["target_pos"].numpy()
-
-    model.eval()
-    predicted_positions, predicted_hd, bottlenecks = model(concat_init, ego_vel)
-
-    data_xy = target_pos.reshape(-1, target_pos.shape[-1])
-    x = data_xy[:, 0]
-    y = data_xy[:, 1]
-
-    activations = bottlenecks.reshape(-1, 256).detach().cpu().numpy()
-    ratemaps = [SAC.calculate_ratemap(x, y, activations[:, kt]) for kt in range(256)]
-
-    scorer = SAC.GridScorer()
-    rated_maps = [scorer.get_grid_score(r) for r in ratemaps]
-
-    return rated_maps
-
-
-def draw_ratemaps(model: nn.Module, device: torch.device, batch: dict):
-    ego_vel = batch["ego_vel"].to(device)
-    encoded_pos = batch["encoded_initial_pos"].to(device)
-    encoded_hd = batch["encoded_initial_hd"].to(device)
-    concat_init = torch.cat([encoded_hd, encoded_pos], axis=2).squeeze()
-    target_pos = batch["target_pos"].numpy()
-
-    model.eval()
-    predicted_positions, predicted_hd, bottlenecks = model(concat_init, ego_vel)
-
-    data_xy = target_pos.reshape(-1, target_pos.shape[-1])
-    x = data_xy[:, 0]
-    y = data_xy[:, 1]
-
-    activations = bottlenecks.reshape(-1, 256).detach().cpu().numpy()
-    ratemaps = [SAC.calculate_ratemap(x, y, activations[:, kt]) for kt in range(256)]
-
-    scorer = SAC.GridScorer()
-    rated_maps = [scorer.get_grid_score(r) for r in ratemaps]
-
-    # Sort by the hexagonal grid factor, descending
-    ratemaps = sorted(rated_maps, key=lambda r: -r.s60)
-
-    fig = validation_views.draw_rated_activations_ratemaps(ratemaps)
-    return fig
-
-
-def review_ratemaps(model_state_path: str):
-    # Run tests on cpu
-    device = torch.device("cpu")
-
-    model = gridcell_models.DeepMindModel()
-    model_state = torch.load(model_state_path)
-    model.load_state_dict(model_state)
-    model.cpu().eval()
-
-    # Make a dataset from a single path that was
-    # not used during trainig and validation
-    paths = glob("data/encoded_pickles/*pickle")
-    path = paths[-1]
-    batch = make_test_batch(path)
-
-    draw_ratemaps(model, device, batch)
